Added file upload controller and moved into controllers package.
